pyproject.toml
[project]
name = "patchscope"
version = "0.1.0"
description = "Multi-agent exploit probability scorer from CVE patch diffs"
requires-python = ">=3.11"
authors = [{name = "Raghu"}]

dependencies = [
    "google-adk>=1.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
markers = [
    "integration: tests that require network access to NVD/GitHub APIs",
]

[tool.ruff]
line-length = 100
target-version = "py311"
